refactor: move cast printutils to common fmt (#2336)
* refactor: move cast printutils to common fmt

* rustmft
